Move GPU RDMA to separate files
